split validation of signatures from fn bodies
